Create Text Mining - Sentence to Vectors using Python
McGill university has conducted an online course to independent students. Using the feedback, grades and demographic data of class participants, we decided to conduct a research to find the characteristics of students who perform well in online classes. This script is used to classify students' answer to a question about their intention for taking the course, into 6 predefined categories that my supervising professor has picked.